minor refactor and cleanup and add docs
